pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple"
extra-index-url = ["https://download.pytorch.org/whl/cu124"]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "aslk"
version = "0.1.0"
description = "AI-Powered Low Thermal Conductivity Material Screening System with Self-Evolution Loop"
authors = [
    {name = "ASLK Team"},
]
dependencies = [
    # Google ADK (Agent Development Kit)
    "google-adk>=1.15.0",
    "agno>=1.8.0",
    "google-generativeai>=0.3.0",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.1.0",
    # Data processing and analysis
    "pandas>=2.1.0",
    "numpy>=1.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    # Chemistry and materials science
    "ase>=3.22.0",
    "pymatgen>=2023.10.0",
    # Machine learning
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "joblib>=1.3.0",
    # PyTorch (with CUDA 12.4 support)
    "torch==2.6.0+cu124",
    "torchvision==0.21.0+cu124",
    "torchaudio==2.6.0+cu124",
    # Database and storage
    "sqlalchemy>=2.0.0",
    "sqlite-utils>=3.35.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "omegaconf>=2.3.0",
    # Logging and monitoring
    "structlog>=23.1.0",
    "rich>=13.6.0",
    "tqdm>=4.66.0",
    # External tools (placeholders - will be integrated later)
    # "crystallm>=0.1.0",  # Structure generation
    # "ai4kappa>=0.1.0",   # Thermal conductivity prediction
    "mattersim>=1.2.0", # Phonon calculation and stability
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}
keywords = ["ai", "materials-science", "thermoelectric", "thermal-conductivity", "agents", "self-evolution"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Chemistry",
    "Topic :: Scientific/Engineering :: Physics",
]

[project.optional-dependencies]
dev = [
    # Development tools
    "black>=23.9.0",
    "isort>=5.12.0",
    "flake8>=6.1.0",
    "mypy>=1.6.0",
    "pre-commit>=3.4.0",
    
    # Testing
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    
    # Documentation
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.4.0",
    "mkdocstrings[python]>=0.23.0",
    
    # Jupyter for analysis
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
]

analysis = [
    # Extended analysis capabilities
    "plotly>=5.17.0",
    "scikit-optimize>=0.9.0",
    "shap>=0.43.0",  # Model interpretability
]

[project.urls]
Homepage = "https://github.com/ASLK-Team/ASLK"
Documentation = "https://aslk.readthedocs.io"
Repository = "https://github.com/ASLK-Team/ASLK.git"
"Bug Tracker" = "https://github.com/ASLK-Team/ASLK/issues"

[project.scripts]
aslk-train = "src.models.train_composition_model:main"
aslk-loop = "src.agents.orchestrator:main"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | data
  | notebooks
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "ase.*",
    "pymatgen.*",
    "xgboost.*",
    "sklearn.*",
    "mattersim.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--tb=short",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/notebooks/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
